Update location, minor tweaks
