+++ library.properties
@@ -1,0 +1,9 @@
+name=CIVmaster
+version=0.2
+author=K7MDL <K7MDL@hotmail.com>
+maintainer=K7MDL <K7MDL@hotmail.com>
+sentence=A CI-V library for the IC-905 and some other Icom CIV radios
+paragraph=An Arduino library forked and extended from https://github.com/WillyIoBrok/CIVmasterLib to support the IC-905
+category=Signal Input/Output
+url=https://github.com/K7MDL2/CIVmasterLib
+architectures=*
